Add commodity group and pdf extractor agents

+++ asklio-procurement-backend/app/models/order_line.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, ForeignKey
+from sqlalchemy import Column, String, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 
@@ -9,6 +9,6 @@ class OrderLine(Base):
     requestID = Column(String, ForeignKey("procurement_request.id"))
     unitPriceCents = Column(Integer, nullable=False)
     unit = Column(String(50), nullable=False)
-    quantity = Column(Integer, nullable=False, default=1)
+    quantity = Column(Float, nullable=False, default=1)
     totalPriceCents = Column(Integer, nullable=False)
     request = relationship("ProcurementRequest", back_populates="order_lines")
